requirements.txt
pygooglenews
python-dotenv
django
djangorestframework
django-cors-headers
